Feature/retrieve money (#3)
withdraw

+++ src/main/java/com/bank/model/Operation.java
@@ -8,10 +8,10 @@ import java.util.Date;
 public class Operation {
     private final String type;
     private final Date date;
-    private final double amount;
-    private final double balance;
+    private final Integer amount;
+    private final Integer balance;
 
-    public Operation(String type, Date date, double amount, double balance) {
+    public Operation(String type, Date date, int amount, int balance) {
         this.type = type;
         this.date = date;
         this.amount = amount;
